+++ src/sora_audio_stream_sink.cpp
@@ -10,6 +10,8 @@
 #include <modules/audio_processing/agc2/rnn_vad/common.h>
 #include <modules/audio_processing/include/audio_frame_view.h>
 
+#include "sora_call.h"
+
 SoraAudioFrameDefaultImpl::SoraAudioFrameDefaultImpl(
     std::unique_ptr<webrtc::AudioFrame> audio_frame)
     : audio_frame_(std::move(audio_frame)) {}
@@ -177,7 +179,7 @@ void SoraAudioStreamSinkImpl::OnData(
     int sample_rate,
     size_t number_of_channels,
     size_t number_of_frames,
-    absl::optional<int64_t> absolute_capture_timestamp_ms) {
+    std::optional<int64_t> absolute_capture_timestamp_ms) {
   auto tuned_frame = std::make_unique<webrtc::AudioFrame>();
   tuned_frame->UpdateFrame(
       0, static_cast<const int16_t*>(audio_data), number_of_frames, sample_rate,
@@ -208,5 +210,6 @@ void SoraAudioStreamSinkImpl::OnData(
     webrtc::RemixFrame(output_channels_, tuned_frame.get());
   }
 
-  on_frame_(std::make_shared<SoraAudioFrame>(std::move(tuned_frame)));
+  call_python(on_frame_,
+              std::make_shared<SoraAudioFrame>(std::move(tuned_frame)));
 }