+++ tests/conftest.py
@@ -22,10 +22,14 @@ def setup():
     if (test_secret_key := os.environ.get("TEST_SECRET_KEY")) is None:
         raise ValueError("TEST_SECRET_KEY is required.")
 
+    if (test_api_url := os.environ.get("TEST_API_URL")) is None:
+        raise ValueError("TEST_API_URL is required.")
+
     return {
         "signaling_urls": test_signaling_urls,
         "channel_id_prefix": test_channel_id_prefix,
         "secret": test_secret_key,
+        "api_url": test_api_url,
         "metadata": {"access_token": test_secret_key},
         # openh264_path は str | None でよい
         "openh264_path": os.environ.get("OPENH264_PATH"),